See desc. (Seriously)
Okay so I've ran into a bit of an issue that's making me question this whole project of doing this in electron. It's difficult to manage the windows of electron and a separate .exe emulator to create a seamless transition when picking a game. I'm trying to hack my way through it but I've got a few ideas if all goes bad:
  - Turn off the screen while it transitions (would REALLY hate to do this)
  - Put game selection on a separate, smaller touchscreen panel (I like this)
  - Use javascript game emulators (really really slow and poor driver support)
  - Use retroarch (wouldn't be mine so it wouldn't feel so good)
Until I decide that a seamless one-screen transition without turning it off is impossible, none of these options will be used.