Initial commit for scraper (see desc.)
My plan is to create an external window accessed by a keyboard binding (so it won't be available in the final product, only "dev mode") that will run through an '/in' folder which contains a load of new roms. It will work through them and scrape the data, sort it into emulator files, add metadata and media. This will make everything much less manual!