Scraper is complete (see desc.)
The scraper is now done but it's not pretty, user friendly and could do with some more validation/error handling. This is not an essential part of the goal since this is the "dev panel" but it would be nice to improve this (which I hope I have time to do).